cliente request terminado

+++ src/LD.Contracts/Requests/WareHouseRequest.cs
@@ -19,7 +19,7 @@ namespace LD.Contracts.Requests
         public string? City { get; set; }
 
         public string? ZipCode { get; set; }
-        public int? Capacity { get; set; }
+        public decimal? Capacity { get; set; }
 
         public bool IsProduction { get; set; } = false;
